requirements.txt
numpy==1.26.4
pandas==2.2.2
pika==1.3.2
python-dotenv==1.0.1
SQLAlchemy==2.0.30
psycopg2==2.9.9
glob2==0.7